test(KIM-382): add @vitest-environment node to 28 test files; split reservations tests
- Add @vitest-environment node directive to all server & API test files for proper Node.js environment isolation
- Split reservations-service.test.ts into reservations-activation.test.ts (21 focused tests) + core service (67 tests)
- Update vitest.config.mts with teardownTimeout: 10000 for reliable test cleanup
- Update HANDOFF.md with complete test suite migration checklist
- Ensure all test files properly configured for Node environment
