+++ __tests__/app/api/tables/activate-route.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { NextRequest, NextResponse } from 'next/server'
 import { ServiceError } from '@/lib/server/service-error'
