+++ __tests__/server/auth-service.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 type ProfileRow = {
