+++ __tests__/server/database-time.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
 const createSupabaseServerAdminClient = vi.fn()
