+++ __tests__/server/member-import.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import { utils, write } from 'xlsx'
