+++ __tests__/server/reservations-service.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import type { SessionUser } from '@/lib/server/auth'
 import { beforeEach, describe, expect, it, vi, afterEach } from 'vitest'
 
@@ -1016,33 +1017,6 @@ describe('reservations service', () => {
       expect(updated.surface).toBeNull()
     })
 
-    it('rejects single-digit hour "9:00" in startTime with 400', async () => {
-      const { updateReservationForSession } = await loadReservationModules()
-
-      await expect(updateReservationForSession(memberSession, 'r1', {
-        startTime: '9:00',
-        endTime: '10:00',
-      })).rejects.toMatchObject({ name: 'ServiceError', statusCode: 400 })
-    })
-
-    it('rejects out-of-range hour "25:00" in endTime with 400', async () => {
-      const { updateReservationForSession } = await loadReservationModules()
-
-      await expect(updateReservationForSession(memberSession, 'r1', {
-        startTime: '16:00',
-        endTime: '25:00',
-      })).rejects.toMatchObject({ name: 'ServiceError', statusCode: 400 })
-    })
-
-    it('rejects out-of-range minutes "18:70" in startTime with 400', async () => {
-      const { updateReservationForSession } = await loadReservationModules()
-
-      await expect(updateReservationForSession(memberSession, 'r1', {
-        startTime: '18:70',
-        endTime: '19:00',
-      })).rejects.toMatchObject({ name: 'ServiceError', statusCode: 400 })
-    })
-
     it('accepts midnight "00:00" as a valid startTime', async () => {
       const { updateReservationForSession } = await loadReservationModules()
 
@@ -1378,693 +1352,6 @@ describe('reservations service', () => {
 
       await expect(checkReservationAccess(memberSession, 'r1')).resolves.toBeUndefined()
       await expect(checkReservationAccess(adminSession, 'r1')).resolves.toBeUndefined()
-    })
-  })
-
-  describe('activateReservationByTable', () => {
-    // Fixed timestamp: 2025-06-15T14:00:00Z (14:00 UTC)
-    const FIXED_DATE = '2025-06-15'
-    const FIXED_TIME = new Date('2025-06-15T14:00:00Z')
-
-    beforeEach(() => {
-      vi.useFakeTimers()
-      vi.setSystemTime(FIXED_TIME)
-    })
-
-    afterEach(() => {
-      vi.useRealTimers()
-      vi.unstubAllEnvs()
-    })
-
-    function seedPendingReservation(overrides?: Partial<ReservationRow>) {
-      const row: ReservationRow = makeReservation({
-        id: 'rA',
-        table_id: 't3',
-        user_id: '2',
-        date: FIXED_DATE,
-        status: 'pending',
-        surface: 'top',
-        ...overrides,
-      })
-      reservationsState.push(row)
-      return row
-    }
-
-    function makeStartTime(offsetMinutes: number): string {
-      const nowUtc = new Date()
-      const clubTz = process.env.CLUB_TIMEZONE ?? 'Europe/Madrid'
-      
-      let madridParts: Record<string, number>
-      try {
-        const parts = new Intl.DateTimeFormat('en-CA', {
-          timeZone: clubTz,
-          year: 'numeric',
-          month: '2-digit',
-          day: '2-digit',
-          hour: '2-digit',
-          minute: '2-digit',
-          second: '2-digit',
-          hour12: false,
-        }).formatToParts(nowUtc)
-        madridParts = Object.fromEntries(
-          parts.filter((p: any) => p.type !== 'literal').map((p: any) => [p.type, parseInt(p.value, 10)])
-        )
-      } catch {
-        // Fallback to UTC if timezone is invalid
-        madridParts = {
-          year: nowUtc.getUTCFullYear(),
-          month: nowUtc.getUTCMonth() + 1,
-          day: nowUtc.getUTCDate(),
-          hour: nowUtc.getUTCHours(),
-          minute: nowUtc.getUTCMinutes(),
-          second: nowUtc.getUTCSeconds(),
-        }
-      }
-      
-      // Create a date in "timezone local time space" (treating parts as UTC for comparison)
-      const madridTime = new Date(
-        madridParts.year!,
-        madridParts.month! - 1,
-        madridParts.day!,
-        madridParts.hour!,
-        madridParts.minute!,
-        madridParts.second!,
-      )
-      madridTime.setMinutes(madridTime.getMinutes() - offsetMinutes)
-      const hh = String(madridTime.getHours()).padStart(2, '0')
-      const mm = String(madridTime.getMinutes()).padStart(2, '0')
-      return `${hh}:${mm}:00`
-    }
-
-    function makeEndTime(startTimeStr: string, durationMinutes: number): string {
-      const [hh, mm] = startTimeStr.split(':')
-      const nowUtc = new Date()
-      const clubTz = process.env.CLUB_TIMEZONE ?? 'Europe/Madrid'
-      
-      let madridParts: Record<string, number>
-      try {
-        const parts = new Intl.DateTimeFormat('en-CA', {
-          timeZone: clubTz,
-          year: 'numeric',
-          month: '2-digit',
-          day: '2-digit',
-          hour12: false,
-        }).formatToParts(nowUtc)
-        madridParts = Object.fromEntries(
-          parts.filter((p: any) => p.type !== 'literal').map((p: any) => [p.type, parseInt(p.value, 10)])
-        )
-      } catch {
-        // Fallback to UTC if timezone is invalid
-        madridParts = {
-          year: nowUtc.getUTCFullYear(),
-          month: nowUtc.getUTCMonth() + 1,
-          day: nowUtc.getUTCDate(),
-        }
-      }
-      
-      const date = new Date(
-        madridParts.year!,
-        madridParts.month! - 1,
-        madridParts.day!,
-        parseInt(hh!, 10),
-        parseInt(mm!, 10),
-        0,
-        0,
-      )
-      date.setTime(date.getTime() + durationMinutes * 60 * 1000)
-      const endHh = String(date.getHours()).padStart(2, '0')
-      const endMm = String(date.getMinutes()).padStart(2, '0')
-      return `${endHh}:${endMm}:00`
-    }
-
-    it('succeeds within the grace period activation window', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      seedPendingReservation({ start_time: makeStartTime(10) })
-
-      const result = await activateReservationByTable('t3', '2', undefined)
-
-      expect(result).toMatchObject({ tableId: 't3', userId: '2', status: 'active' })
-    })
-
-    it('throws CHECK_IN_TOO_EARLY when called before start_time', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      seedPendingReservation({ start_time: makeStartTime(-30) })
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 400,
-        message: expect.stringContaining('CHECK_IN_TOO_EARLY'),
-      })
-    })
-
-    it('boundary: called exactly 5 minutes before start_time (early window opens) succeeds', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, start_time is 14:05:00 (5 min in future)
-      // This is exactly at [start - 5min, end_time], so check-in should succeed
-      const startTime = makeStartTime(-5)
-      seedPendingReservation({ start_time: startTime })
-
-      const result = await activateReservationByTable('t3', '2', undefined)
-
-      expect(result.status).toBe('active')
-    })
-
-    it('boundary: called 6 minutes before start_time throws CHECK_IN_TOO_EARLY', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, start_time is 14:06:00 (6 min in future)
-      // This is before the early window [start - 5min], so check-in should fail
-      seedPendingReservation({ start_time: makeStartTime(-6) })
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 400,
-        message: expect.stringContaining('CHECK_IN_TOO_EARLY'),
-      })
-    })
-
-    it('throws CHECK_IN_TOO_LATE when called after end_time', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      const startTime = makeStartTime(25)
-      seedPendingReservation({ start_time: startTime, end_time: makeEndTime(startTime, 20) })
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 400,
-        message: expect.stringContaining('CHECK_IN_TOO_LATE'),
-      })
-    })
-
-    it('throws CHECK_IN_ALREADY_ACTIVE when reservation is already active', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      reservationsState.push(makeReservation({
-        id: 'rActive',
-        table_id: 't3',
-        user_id: '2',
-        date: FIXED_DATE,
-        status: 'active',
-        surface: 'top',
-        start_time: makeStartTime(10),
-      }))
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 409,
-        message: expect.stringContaining('CHECK_IN_ALREADY_ACTIVE'),
-      })
-    })
-
-    it('throws CHECK_IN_NO_RESERVATION when no pending reservation exists', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 404,
-        message: expect.stringContaining('CHECK_IN_NO_RESERVATION'),
-      })
-    })
-
-    it('with side=inf: matches reservation with surface bottom', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      seedPendingReservation({ surface: 'bottom', start_time: makeStartTime(10) })
-
-      const result = await activateReservationByTable('t3', '2', 'inf')
-
-      expect(result).toMatchObject({ tableId: 't3', userId: '2', status: 'active' })
-    })
-
-    it('with side=inf on a non-removable-top table: throws CHECK_IN_NO_RESERVATION', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      reservationsState.push(makeReservation({
-        id: 'rNonTop',
-        table_id: 't1',
-        user_id: '2',
-        date: FIXED_DATE,
-        status: 'pending',
-        surface: null,
-        start_time: makeStartTime(10),
-      }))
-
-      await expect(activateReservationByTable('t1', '2', 'inf')).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 404,
-        message: expect.stringContaining('CHECK_IN_NO_RESERVATION'),
-      })
-    })
-
-    it('boundary: called exactly at start_time succeeds', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      seedPendingReservation({ start_time: makeStartTime(0) })
-
-      const result = await activateReservationByTable('t3', '2', undefined)
-
-      expect(result.status).toBe('active')
-    })
-
-    it('boundary: called at start_time + (19) min succeeds', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      seedPendingReservation({ start_time: makeStartTime(19) })
-
-      const result = await activateReservationByTable('t3', '2', undefined)
-
-      expect(result.status).toBe('active')
-    })
-
-    it('boundary: called at start_time + 30 min succeeds', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, start_time is 13:30:00 (30 min in past)
-      // Within 60-min window: should succeed
-      seedPendingReservation({ start_time: makeStartTime(30) })
-
-      const result = await activateReservationByTable('t3', '2', undefined)
-
-      expect(result.status).toBe('active')
-    })
-
-    it('boundary: called at start_time + 60 min succeeds when reservation longer', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, start_time is 13:00:00 (60 min in past)
-      // Exactly at 60-min boundary, end_time is 14:30:00 (extends past boundary)
-      // windowEnd = min(start + 60min, end) = min(14:00:00, 14:30:00) = 14:00:00
-      // now === windowEnd → should succeed (not strictly greater)
-      const startTime = makeStartTime(60)
-      const endTime = makeEndTime(startTime, 90)
-      seedPendingReservation({ start_time: startTime, end_time: endTime })
-
-      const result = await activateReservationByTable('t3', '2', undefined)
-
-      expect(result.status).toBe('active')
-    })
-
-    it('boundary: called at start_time + 61 min throws CHECK_IN_TOO_LATE', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, start_time is 12:59:00 (61 min in past)
-      // Beyond 60-min window: should fail
-      seedPendingReservation({ start_time: makeStartTime(61) })
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 400,
-        message: expect.stringContaining('CHECK_IN_TOO_LATE'),
-      })
-    })
-
-    it('boundary: end_time before 60-min mark caps window at end_time', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, start_time is 13:30:00 (30 min in past)
-      // end_time is 13:40:00 (20 min in past, before 60-min window closes)
-      // windowEnd = min(start + 60min, end) = min(14:30:00, 13:40:00) = 13:40:00
-      // now > windowEnd → should fail
-      const startTime = makeStartTime(30)
-      const endTime = makeEndTime(startTime, 10)
-      seedPendingReservation({ start_time: startTime, end_time: endTime })
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 400,
-        message: expect.stringContaining('CHECK_IN_TOO_LATE'),
-      })
-    })
-
-    it('boundary: called exactly at end_time succeeds', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, set start_time so end_time is also 14:00:00
-      // At end_time boundary, check-in should succeed (now === reservationEnd, not strictly greater)
-      const startTime = makeStartTime(20)
-      const endTime = makeEndTime(startTime, 20)
-      seedPendingReservation({ start_time: startTime, end_time: endTime })
-
-      const result = await activateReservationByTable('t3', '2', undefined)
-
-      expect(result.status).toBe('active')
-    })
-
-    it('boundary: called just after end_time throws CHECK_IN_TOO_LATE', async () => {
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // now is 14:00:00, set end_time to 13:59:00 (1 minute before now)
-      // This is after end_time, so check-in should fail
-      const startTime = makeStartTime(21)
-      const endTime = makeEndTime(startTime, 19)
-      seedPendingReservation({ start_time: startTime, end_time: endTime })
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 400,
-        message: expect.stringContaining('CHECK_IN_TOO_LATE'),
-      })
-    })
-
-    it('non-removable-top (large) table: activates with surface=null without surface filter', async () => {
-      // Task 2: happy-path check-in on a non-removable-top table
-      // t1 is type 'large' — isRemovableTop=false → no surface filter applied
-      const { activateReservationByTable } = await loadReservationModules()
-
-      reservationsState.push(makeReservation({
-        id: 'rLarge',
-        table_id: 't1',
-        user_id: '2',
-        date: FIXED_DATE,
-        status: 'pending',
-        surface: null,
-        start_time: makeStartTime(10),
-      }))
-
-      const result = await activateReservationByTable('t1', '2', undefined)
-
-      expect(result).toMatchObject({ tableId: 't1', userId: '2', status: 'active' })
-    })
-
-    it('activeQuery DB error returns 500', async () => {
-      // Task 3: when the active-reservation query errors, the service should throw 500.
-      // The pendingQuery maybeSingle returns null (no row, no error) → service moves on to activeQuery.
-      // The activeQuery maybeSingle returns an error → service should throw 500.
-      // We use vi.doMock here and restore the original mock factory at the end of the test
-      // to prevent contamination of subsequent tests.
-      vi.resetModules()
-
-      // Track how many times maybeSingle has been called across all chains in this test
-      let maybeSingleCallCount = 0
-
-      vi.doMock('@/lib/supabase/server', () => ({
-        createSupabaseServerClient: vi.fn(async () => ({
-          from: vi.fn((table: string) => {
-            if (table === 'tables') {
-              return {
-                select: vi.fn(() => ({
-                  eq: vi.fn((column: string, value: string) => ({
-                    maybeSingle: vi.fn(async () => ({
-                      data: column === 'id' ? (tablesState.get(value) ?? null) : null,
-                      error: null,
-                    })),
-                  })),
-                })),
-              }
-            }
-            return { select: vi.fn(() => buildSelectChain(reservationsState)) }
-          }),
-        })),
-        createSupabaseServerAdminClient: vi.fn(() => {
-          function makeChain(): {
-            eq: (col: string, val: string) => ReturnType<typeof makeChain>
-            maybeSingle: () => Promise<{ data: null; error: { message: string } | null }>
-            select: (cols?: string) => ReturnType<typeof makeChain>
-            update: (payload: unknown) => ReturnType<typeof makeChain>
-            single: () => Promise<{ data: null; error: null }>
-          } {
-            return {
-              eq: vi.fn(() => makeChain()),
-              select: vi.fn(() => makeChain()),
-              update: vi.fn(() => makeChain()),
-              single: vi.fn(async () => ({ data: null, error: null })),
-              maybeSingle: vi.fn(async () => {
-                maybeSingleCallCount++
-                if (maybeSingleCallCount === 1) {
-                  // First call: pendingQuery — no pending reservation found
-                  return { data: null, error: null }
-                }
-                // Second call: activeQuery — simulate DB error
-                return { data: null, error: { message: 'db error' } }
-              }),
-            }
-          }
-
-          return {
-            from: vi.fn(() => makeChain()),
-            rpc: vi.fn(),
-          }
-        }),
-      }))
-
-      const { activateReservationByTable: activate } = await import('@/lib/server/reservations-service')
-
-      await expect(activate('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 500,
-      })
-
-      // Restore the original mock factory so subsequent tests are not contaminated
-      vi.resetModules()
-      vi.doMock('@/lib/supabase/server', () => ({
-        createSupabaseServerClient: vi.fn(async () => ({
-          from: vi.fn((table: string) => {
-            if (table === 'tables') {
-              return {
-                select: vi.fn(() => ({
-                  eq: vi.fn((column: string, value: string) => ({
-                    maybeSingle: vi.fn(async () => ({
-                      data: column === 'id' ? (tablesState.get(value) ?? null) : null,
-                      error: null,
-                    })),
-                  })),
-                })),
-              }
-            }
-            if (table === 'reservations') {
-              return {
-                select: vi.fn(() => buildSelectChain(reservationsState)),
-                insert: vi.fn((payload: Record<string, string | null>) => ({
-                  select: vi.fn(() => ({
-                    single: vi.fn(async () => {
-                      const row = makeReservation({
-                        id: `r${reservationsState.length + 1}`,
-                        table_id: String(payload.table_id),
-                        user_id: String(payload.user_id),
-                        date: String(payload.date),
-                        start_time: `${String(payload.start_time)}:00`,
-                        end_time: `${String(payload.end_time)}:00`,
-                        surface: payload.surface as ReservationRow['surface'],
-                        created_at: '2026-04-04T12:00:00.000Z',
-                      })
-                      reservationsState.push(row)
-                      return { data: cloneReservation(row), error: null }
-                    }),
-                  })),
-                })),
-                update: vi.fn((payload: Record<string, string | null>) => ({
-                  eq: vi.fn((column: string, value: string) => ({
-                    select: vi.fn(() => ({
-                      single: vi.fn(async () => {
-                        const row = reservationsState.find((reservation) => String((reservation as Record<string, unknown>)[column]) === value)
-                        if (!row) {
-                          return { data: null, error: null }
-                        }
-                        Object.assign(row, payload)
-                        if (payload.start_time) row.start_time = `${payload.start_time}:00`
-                        if (payload.end_time) row.end_time = `${payload.end_time}:00`
-                        return { data: cloneReservation(row), error: null }
-                      }),
-                    })),
-                  })),
-                })),
-              }
-            }
-            throw new Error(`Unexpected table ${table}`)
-          }),
-        })),
-        createSupabaseServerAdminClient: vi.fn(() => ({
-          from: vi.fn((table: string) => {
-            if (table !== 'reservations') {
-              throw new Error(`Unexpected admin table ${table}`)
-            }
-            return {
-              select: vi.fn(() => buildSelectChain(reservationsState)),
-              update: vi.fn((payload: Record<string, unknown>) => {
-                const filters: Array<[string, string]> = []
-                const chain = {
-                  eq: vi.fn((column: string, value: string) => {
-                    filters.push([column, value])
-                    return chain
-                  }),
-                  select: vi.fn(() => ({
-                    single: vi.fn(async () => {
-                      const row = reservationsState.find((r) =>
-                        filters.every(([col, val]) => String((r as Record<string, unknown>)[col]) === val)
-                      )
-                      if (!row) {
-                        return { data: null, error: null }
-                      }
-                      Object.assign(row, payload)
-                      return { data: cloneReservation(row), error: null }
-                    }),
-                  })),
-                }
-                return chain
-              }),
-            }
-          }),
-          rpc: vi.fn(),
-        })),
-      }))
-    })
-
-    it('invalid CLUB_TIMEZONE propagates as a RangeError', async () => {
-      // Task 4: CLUB_TIMEZONE='Invalid/Zone' → RangeError propagates (no fallback)
-      vi.stubEnv('CLUB_TIMEZONE', 'Invalid/Zone')
-      const { activateReservationByTable } = await loadReservationModules()
-
-      seedPendingReservation({ start_time: makeStartTime(10) })
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toThrow(RangeError)
-    })
-
-    it('getTable returning null throws 404 with Table not found', async () => {
-      // Task 5: when the table does not exist in the DB, service throws 404
-      const { activateReservationByTable } = await loadReservationModules()
-
-      // Use a tableId that is not seeded in tablesState
-      await expect(activateReservationByTable('t-nonexistent', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 404,
-        message: expect.stringContaining('Table not found'),
-      })
-    })
-
-    it('UPDATE returning PGRST116 resolves to 409 CHECK_IN_ALREADY_ACTIVE (TOCTOU race)', async () => {
-      // Simulates the TOCTOU race: pendingQuery finds the row, but by the time the UPDATE
-      // executes the row is gone (already activated by a concurrent request).
-      // PostgREST returns PGRST116 when .single() matches zero rows.
-      // We override createSupabaseServerAdminClient for this one invocation via mockReturnValueOnce.
-      const { createSupabaseServerAdminClient } = await import('@/lib/supabase/server')
-      const adminMock = vi.mocked(createSupabaseServerAdminClient)
-
-      const pendingRow = makeReservation({
-        id: 'rTOCTOU',
-        table_id: 't3',
-        user_id: '2',
-        date: FIXED_DATE,
-        status: 'pending',
-        surface: 'top',
-        start_time: makeStartTime(10),
-      })
-
-      // Build a select chain that returns the pending row for pendingQuery
-      function buildSingleRowSelectChain(row: ReservationRow) {
-        const self: {
-          eq: (col: string, val: string) => typeof self
-          maybeSingle: () => Promise<{ data: ReservationRow; error: null }>
-        } = {
-          eq: vi.fn(() => self),
-          maybeSingle: vi.fn(async () => ({ data: { ...row }, error: null })),
-        }
-        return self
-      }
-
-      // Build an update chain whose single() returns PGRST116
-      function buildPGRST116UpdateChain() {
-        const self: {
-          eq: (col: string, val: string) => typeof self
-          select: (cols?: string) => typeof self
-          single: () => Promise<{ data: null; error: { code: string; message: string } }>
-        } = {
-          eq: vi.fn(() => self),
-          select: vi.fn(() => self),
-          single: vi.fn(async () => ({
-            data: null,
-            error: { code: 'PGRST116', message: 'JSON object requested, multiple (or no) rows returned' },
-          })),
-        }
-        return self
-      }
-
-      let fromCallCount = 0
-      adminMock.mockReturnValueOnce({
-        from: vi.fn((table: string) => {
-          if (table !== 'reservations') throw new Error(`Unexpected admin table ${table}`)
-          fromCallCount++
-          if (fromCallCount === 1) {
-            // pendingQuery: .select(cols).eq().eq().eq().eq().eq().maybeSingle()
-            return { select: vi.fn(() => buildSingleRowSelectChain(pendingRow)) }
-          }
-          // updateChain: .update().eq().eq().select().single() → PGRST116
-          return { update: vi.fn(() => buildPGRST116UpdateChain()) }
-        }),
-        rpc: createDatabaseTimeRpc(),
-      } as unknown as ReturnType<typeof createSupabaseServerAdminClient>)
-
-      const { activateReservationByTable } = await loadReservationModules()
-
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 409,
-        message: expect.stringContaining('CHECK_IN_ALREADY_ACTIVE'),
-      })
-    })
-
-    it('sequential double-call returns 409 CHECK_IN_ALREADY_ACTIVE on second call', async () => {
-      // First call activates the reservation (pending → active).
-      // Second call finds no pending row but finds an active row → 409 CHECK_IN_ALREADY_ACTIVE.
-      const { createSupabaseServerAdminClient } = await import('@/lib/supabase/server')
-      const adminMock = vi.mocked(createSupabaseServerAdminClient)
-      adminMock.mockImplementation(() => ({
-        from: vi.fn((table: string) => {
-          if (table === 'event_room_blocks') {
-            return {
-              select: vi.fn(() => buildSelectChain(eventRoomBlocksState)),
-            }
-          }
-
-          if (table !== 'reservations') {
-            throw new Error(`Unexpected admin table ${table}`)
-          }
-
-          return {
-            select: vi.fn(() => buildSelectChain(reservationsState)),
-            update: vi.fn((payload: Record<string, unknown>) => {
-              const filters: Array<[string, string]> = []
-              const chain = {
-                eq: vi.fn((column: string, value: string) => {
-                  filters.push([column, value])
-                  return chain
-                }),
-                select: vi.fn(() => ({
-                  single: vi.fn(async () => {
-                    const row = reservationsState.find((reservation) =>
-                      filters.every(([column, value]) => String((reservation as Record<string, unknown>)[column]) === value)
-                    )
-                    if (!row) {
-                      return { data: null, error: null }
-                    }
-                    Object.assign(row, payload)
-                    return { data: cloneReservation(row), error: null }
-                  }),
-                })),
-              }
-              return chain
-            }),
-          }
-        }),
-        rpc: createDatabaseTimeRpc(),
-      }) as unknown as ReturnType<typeof createSupabaseServerAdminClient>)
-
-      const { activateReservationByTable } = await loadReservationModules()
-
-      seedPendingReservation({ start_time: makeStartTime(10) })
-
-      // First activation should succeed
-      const firstResult = await activateReservationByTable('t3', '2', undefined)
-      expect(firstResult).toMatchObject({ tableId: 't3', userId: '2', status: 'active' })
-
-      // Second activation on the same table/user — reservation is now active, not pending
-      await expect(activateReservationByTable('t3', '2', undefined)).rejects.toMatchObject({
-        name: 'ServiceError',
-        statusCode: 409,
-        message: expect.stringContaining('CHECK_IN_ALREADY_ACTIVE'),
-      })
     })
   })
