+++ __tests__/server/rooms-service.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 import type { ServiceError } from '@/lib/server/service-error'
