+++ __tests__/utils/date.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import { describe, it, expect } from 'vitest'
 import { formatDate, formatTime, generateTimeSlots } from '@/lib/utils'
