+++ __tests__/utils/password.test.ts
@@ -1,3 +1,4 @@
+// @vitest-environment node
 import { describe, it, expect } from 'vitest'
 import { validatePassword } from '@/lib/validations/password'
