+++ package.json
@@ -11,9 +11,9 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "test": "NODE_OPTIONS=--max-old-space-size=4096 vitest run",
-    "test:watch": "NODE_OPTIONS=--max-old-space-size=4096 vitest",
-    "test:coverage": "NODE_OPTIONS=--max-old-space-size=4096 vitest run --coverage",
+    "test": "vitest run",
+    "test:watch": "vitest",
+    "test:coverage": "vitest run --coverage",
     "test:integration": "node -e \"if (process.platform === 'win32') { console.log('Skipping integration shell check on Windows.'); process.exit(0); } require('node:child_process').execFileSync('bash', ['./scripts/supabase-integration-check.sh'], { stdio: 'inherit' });\"",
     "lint": "next lint --no-cache",
     "typecheck": "next typegen && node -e \"require('node:fs').rmSync('.next/cache/.tsbuildinfo', { force: true })\" && tsc --noEmit",
